Move Json operations out of pg_ops
